refactor(editor): split the 1975-line onlyoffice-editor into a facade and one module per concern
lib/onlyoffice-editor.ts had grown to seven unrelated concerns in one file, and
prepareEditorIframe alone was 584 lines: nine independent vendor patches inside
a single frame loop, each with its own `win as unknown as {...}` cast and
`__oo*` flag, with a "everything is in place" check reading five locals
scattered through it.

It is now a facade -- mount, rebuild, loadEditorApi -- over lib/onlyoffice/:
iframe-guards (orchestration) with one file per guard under guards/, open-state,
open-failure, font-system, save-stream, viewport, sdk-api, readonly, ui-theme,
file-helpers. Each guard file leads with the production defect it exists for.

The public surface is unchanged: the facade re-exports every symbol
lib/embed-api, lib/web-mcp, lib/converter and the unit tests import, so not a
line of test code moves.

Two rules the split follows:
- Cross-module state has exactly one owner. documentContentReady /
  documentOpenError / lastFrameError were module-level variables read and
  written from three places; they now live in open-state and everyone else
  calls its accessors. Readonly belongs to readonly.ts, the in-flight save
  request to save-stream.ts (the guard fails it through
  failPendingSaveConversion instead of reaching into another module's state).
  Duplicating any of it is the one thing that would actually introduce bugs.
- No cycles. The retry path needs createEditorInstance, which lives in the
  facade, so it goes through an injected runner (setOpenRunner) the same way
  the converter and UI callbacks already do.

Behaviour is unchanged by construction: every line of the old file was matched
verbatim against the new tree, and the 39 differences are all deliberate
(added `export`, module variables read through accessors, the injected runner,
moved imports). Guards report "in place" as a return value; iframe-guards
combines the same five as before.

Verified: tsc + oxlint clean, 617 unit tests green (untouched), full local e2e
against the real editor 82 passed / 16 skipped / 0 failed.
